Updated again

+++ Components/Division.cs
@@ -3,7 +3,7 @@ namespace Components.Division
     
     class Division
     {
-        public static float Devide(int a, int b)
+        public static int Devide(int a, int b)
         {
             return (a / b);
         }
